Undo-Redo Function added to the Editor
A class named History was added. It contains data structures that store
recent states of the canvas. A snapshot of the editor is stored whenever
the mouse state is GLUT_UP and GLUT_LEFT_BUTTON. The F2 and F3 special
function keys are used to undo and redo actions respectively.

Satvik's favourite picture was also added. :-P