Flood-Fill Function partially fixed; Unfilled Rect Tool added
Due to the usage of GL_POINT instead of GL_POINTS, the flood fill
function wasn't drawing anything on the canvas. Now that the error has
been discovered, flood filling works. Unfortunately, it is too slow and
the depth has to be limited to 200.

The Circle tool now constructs circles with lines, instead of points, in
order to aid the flood-fill algorithm.

The Cone-Drawing Tool was converted to an Unfilled Rectangle-Drawing
Tool.