The translate tool is now more user-friendly
The user can now adjust the size of the selection box as well as the
location at which to paste the selection, using mouse motion.

The Rotate button was replaced with a Cut Button, and the translate tool
will be known as a cut-paste tool. The cut-paste and copy-paste
operators were removed from the Menu Bar.