Text Tool added
Now, the friendly-user can add strings of characters to the canvas. Just
because C++ doesn't support cyclic dependencies, the tool bar cannot
create an object of the Typewriter class. So, code had to be added for
the text tool to communicate with the menu bar and request it for its
typewriter.