Scaling Tool is finally working!
The scaling tool is derived from the translate class, and uses the same
selection box logic through a callback function. Support for callback
functions in C++ was added, thanks to an intuitive solution which falls
within the realms of templates and compile time instantiation and
specialization, from
http://www.partow.net/programming/templatecallback/index.html#TheSolution
. Using the power of this class, code size was reduced drastically.

Now, I guess all rubberbanding tools are running as fast as possible
with the gl logic operations.