Flood fill tool is finally working
After going crazy, we have finally returned. The flood fill tool is
finally in working condition and runs very quickly, thanks to a
queue-based implementation. A 3-D array was used to store color values
of the pixels of the canvas. The recursive implementation kept on
causing stack-overflows. That's why a queueFill() method was created
along with a Node class. There is still a small bug - When the tool is
used on an empty canvas, a memory exception occurs.

The implementations that didn't work at all, have also been retained in
our program, just for old times' sake, to remember how the flood fill
algorithm tortured us.