Alpha (Transparency) values can be changed using the F4 and F5 keys
Translucent, Transparent and Opaque figures can be drawn, using the F4
and F5 keys to increase or decrease alpha values. This editor can now be
used to produce amazing art. In addition to this, the MenuBar was
updated to display some helpful commands to the user.

This is the last update that Project Codename: Chicken Curry 7x will
see. Goodbye, Chicken Curry. It was nice knowing you. :-)